+++ src/Schema/Directives/CacheKeyDirective.php
@@ -14,7 +14,7 @@ class CacheKeyDirective extends BaseDirective implements Directive
 """
 Specify the field to use as a key when creating a cache.
 """
-directive @cacheKey on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
+directive @cacheKey on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION | FIELD_DEFINITION
 GRAPHQL;
     }
 }