+++ src/Testing/UsesTestSchema.php
@@ -2,6 +2,7 @@
 
 namespace Nuwave\Lighthouse\Testing;
 
+use Exception;
 use Nuwave\Lighthouse\Schema\Source\SchemaSourceProvider;
 
 trait UsesTestSchema
@@ -19,7 +20,7 @@ trait UsesTestSchema
             SchemaSourceProvider::class,
             function (): TestSchemaProvider {
                 if (! isset($this->schema)) {
-                    throw new \Exception('Missing test schema, provide one by setting $this->schema.');
+                    throw new Exception('Missing test schema, provide one by setting $this->schema.');
                 }
 
                 return new TestSchemaProvider($this->schema);